fix syncbn, test=develop
